requirements.txt
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.19.1
fsspec==2025.9.0
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyclipper==1.3.0.post6
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.16.2
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.9.9
tk==0.1.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
